refactor: distribute styles
