+++ src/core/components/Canvas/styles/CanvasWrapper.ts
