+++ src/core/components/Canvas/styles/ControlsWrapper.ts
