+++ src/core/components/Canvas/styles/StyledBtnsInEditor.ts
