+++ src/core/components/Canvas/styles/StyledCanvasBtns.ts
