+++ src/core/components/Canvas/styles/StyledCommonCanvas.ts
