+++ src/core/components/Canvas/styles/StyledGallery.ts
