+++ src/core/components/Canvas/styles/StyledGalleryWrapper.ts
