+++ src/core/components/Header/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { ToastContainer } from 'react-toastify';
-import HeaderWrapper from '../styles/HeaderWrapper';
-import StyledAvatar from '../styles/StyledAvatar';
-import StyledButton from '../styles/StyledButton';
-import StyledFlexRow from '../styles/StyledFlexRow';
-import StyledTitle from '../styles/StyledTitle';
+import HeaderWrapper from './styles/HeaderWrapper';
 import { logOutAction } from '../../actions/authActions';
 import { AppState } from '../../interfaces';
+import StyledButton from '../styles/buttons/StyledButton';
+import StyledAvatar from '../styles/common/StyledAvatar';
+import StyledFlexRow from '../styles/common/StyledFlexRow';
+import StyledTitle from '../styles/common/StyledTitle';
 
 const Header: React.FC = () => {
   const authenticated = useSelector((state: AppState) => state.auth.authenticated);
