+++ src/core/components/styles/buttons/StyledDeleteBtn.ts
