+++ src/core/components/styles/common/EditorWrapper.ts
