+++ src/core/components/styles/common/StyledApp.ts
