+++ src/core/components/styles/common/StyledAvatar.ts
