+++ src/core/components/styles/common/StyledContainer.ts
