+++ src/core/components/styles/common/StyledTitle.ts
