+++ src/core/components/styles/forms/StyledOption.ts
