+++ src/core/components/styles/modalWindow/StyledModalBg.ts
