+++ src/core/components/styles/modalWindow/StyledModalButton.ts
