+++ src/core/components/styles/modalWindow/StyledModalWindow.ts
