+++ src/pages/register/index.tsx
@@ -4,11 +4,11 @@ import { Link } from 'react-router-dom';
 import 'react-toastify/dist/ReactToastify.css';
 import { createUserAction } from '../../core/actions/authActions';
 import Input from '../../core/components/Input';
-import StyledButton from '../../core/components/styles/StyledButton';
-import StyledContainer from '../../core/components/styles/StyledContainer';
-import StyledForm from '../../core/components/styles/StyledForm';
-import StyledLinkDiv from '../../core/components/styles/StyledLinkDiv';
-import StyledTitle from '../../core/components/styles/StyledTitle';
+import StyledButton from '../../core/components/styles/buttons/StyledButton';
+import StyledContainer from '../../core/components/styles/common/StyledContainer';
+import StyledTitle from '../../core/components/styles/common/StyledTitle';
+import StyledForm from '../../core/components/styles/forms/StyledForm';
+import StyledLinkDiv from '../../core/components/styles/common/StyledLinkDiv';
 import RoutesConst from '../../core/constants/routesConst';
 import notify from '../../core/helpers/notify';
 import { AppState } from '../../core/interfaces';
